Added a new file that keeps track of pain messages, along with other small enhancements.
Updated all of the add and remove dialog messages to notify the user of the files that have something added or removed from it in the scenario editor menu.
You can now view how many entries their are for the pains file in the statistics menu.
Added the ability to add, remove, and view pain messages in the scenario editor menu.
Added a new file called pains.list. This file will be discussed more in depth in the readme.