Added the ability to create non hurtfull scenarios.
Added the ability to add, edit, remove, and view effect statuses in the scenario manager menu. You can also view how many entries their are for the effects file in the scenario statistics menu.
Added a new file called effects.list. This file will be discussed more in depth in the readme.